+++ src/spead2_recv.cpp
@@ -1,4 +1,4 @@
-/* Copyright 2015, 2018, 2019 SKA South Africa
+/* Copyright 2015, 2018-2020 SKA South Africa
  *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -106,7 +106,7 @@ static options parse_args(int argc, const char **argv)
         ("pyspead", make_opt(opts.pyspead), "Be bug-compatible with PySPEAD")
         ("joint", make_opt(opts.joint), "Treat all sources as a single stream")
         ("tcp", make_opt(opts.tcp), "Receive data over TCP instead of UDP")
-        ("bind", make_opt(opts.bind), "Interface address for multicast")
+        ("bind", make_opt(opts.bind), "Interface address")
         ("packet", make_opt_no_default(opts.packet), "Maximum packet size to use")
         ("buffer", make_opt_no_default(opts.buffer), "Socket buffer size")
         ("threads", make_opt(opts.threads), "Number of worker threads")
@@ -348,7 +348,9 @@ static std::unique_ptr<spead2::recv::stream> make_stream(
         else
         {
             udp::resolver resolver(thread_pool.get_io_service());
-            udp::resolver::query query(host, port);
+            udp::resolver::query query(host, port,
+                boost::asio::ip::udp::resolver::query::address_configured
+                | boost::asio::ip::udp::resolver::query::passive);
             udp::endpoint endpoint = *resolver.resolve(query);
 #if SPEAD2_USE_IBV
             if (opts.ibv)
@@ -357,8 +359,7 @@ static std::unique_ptr<spead2::recv::stream> make_stream(
             }
             else
 #endif
-            if (endpoint.address().is_multicast() && endpoint.address().is_v4()
-                && !opts.bind.empty())
+            if (endpoint.address().is_v4() && !opts.bind.empty())
             {
                 stream->emplace_reader<spead2::recv::udp_reader>(
                     endpoint, opts.packet, opts.buffer,
@@ -367,7 +368,7 @@ static std::unique_ptr<spead2::recv::stream> make_stream(
             else
             {
                 if (!opts.bind.empty())
-                    std::cerr << "--bind is only applicable to IPv4 multicast, ignoring\n";
+                    std::cerr << "--bind is not implemented for IPv6\n";
                 stream->emplace_reader<spead2::recv::udp_reader>(endpoint, opts.packet, opts.buffer);
             }
         }